Re-factor judge case assignment logic into its own class and move task validation into a shared concern
