+++ app/models/attorney_case_review.rb
@@ -1,4 +1,6 @@
 class AttorneyCaseReview < ApplicationRecord
+  include LegacyTaskConcern
+
   belongs_to :reviewing_judge, class_name: "User"
   belongs_to :attorney, class_name: "User"
 
@@ -13,15 +15,13 @@ class AttorneyCaseReview < ApplicationRecord
 
   attr_accessor :issues
 
-  # task ID is vacols_id concatenated with the date assigned
-  validates :task_id, format: { with: /\A[0-9A-Z]+-[0-9]{4}-[0-9]{2}-[0-9]{2}\Z/i }
-
   def appeal
     @appeal ||= LegacyAppeal.find_or_create_by(vacols_id: vacols_id)
   end
 
   def reassign_case_to_judge_in_vacols!
     attorney.access_to_task?(vacols_id)
+
     AttorneyCaseReview.repository.reassign_case_to_judge!(
       vacols_id: vacols_id,
       created_in_vacols_date: created_in_vacols_date,
@@ -54,14 +54,6 @@ class AttorneyCaseReview < ApplicationRecord
   end
 
   private
-
-  def vacols_id
-    task_id.split("-", 2).first
-  end
-
-  def created_in_vacols_date
-    task_id.split("-", 2).second.to_date
-  end
 
   class << self
     attr_writer :repository
