--- app/models/queues/attorney_queue.rb
@@ -1,9 +1,0 @@
-class AttorneyQueue < WorkQueue
-  # We don't redefine the repository as a fake
-  # on BaseQueue until after autoloading happens,
-  # so if we rely on the autoloaded method we'll never
-  # get the fake. Awwwkward.
-  def self.repository
-    WorkQueue.repository
-  end
-end
